+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-08-11 20:49:35
-// Commit: test(network): implement widget integration tests for network
+// Updated at: 2026-08-13 13:22:52
+// Commit: refactor(network): extract reusable atomic widgets from network
 
 class networkComponentState {
-  final String commitRef = '3547';
-  final String lastMessage = 'test(network): implement widget integration tests for network';
-  final DateTime timestamp = DateTime.parse('2026-08-11T20:49:35+05:00');
+  final String commitRef = '3573';
+  final String lastMessage = 'refactor(network): extract reusable atomic widgets from network';
+  final DateTime timestamp = DateTime.parse('2026-08-13T13:22:52+05:00');
 }